fix panic on unknown code_page
cp_to_string() panics when the code page is unknown, but some documents are
ascii only with a code_page of 0.

When the code_page is unknown, default to UTF-8 instead of panicking. If we guess
wrong, decode_to_string() will replace invalid byte sequences with the
REPLACEMENT CHARACTER anyway.