fix panic in decompression of CopyToken
A decompression CopyToken specifies a range of bytes from the `result` buffer
to copy back at the end. We have to check that the range is valid, otherwise
we will panic when indexing.

Note that the range can span bytes past the `result` cursor that
are not yet written at the start of the copy, this is why we can't use
Vec::extend_from_within().